requirements.txt
# pip install -r requirements.txt 
Pillow>=8.1.0
